ImageProcessor/requirements.txt
requests
Pillow
pillow-heif